Remove unused generator code
