+++ SymbolsGenerator/Sources/SymbolsGenerator/Helpers/CollectionExtension.swift
@@ -8,3 +8,14 @@ extension Sequence {
         return sorted { areInIncreasingOrder(element($0), element($1)) }
     }
 }
+
+infix operator ×
+internal func ×<A, B>(lhs: [A], rhs: [B]) -> [(A, B)] {
+    var result = [(A, B)]()
+    for a in lhs {
+        for b in rhs {
+            result.append((a, b))
+        }
+    }
+    return result
+}
