+++ SymbolsGenerator/Sources/SymbolsGenerator/Helpers/Types.swift
@@ -16,15 +16,9 @@ struct Symbol {
     var availableLayersets: [Availability: Set<String>]
     var olderSymbol: ScannedSymbol?
     var newerSymbol: ScannedSymbol?
-//    var type: SymbolType
+
     var propertyName: String { name.toPropertyName }
 }
-
-//enum SymbolType {
-//  case normal
-//  case replacement(for: ScannedSymbol)
-//  case replaced(by: ScannedSymbol)
-//}
 
 struct Availability: Comparable, Equatable, Hashable {
     var iOS: String
